🚑️chore: AricleBoards 관련 경로 수정
