+++ components/Boards/BestArticleBoards/BestArticleList.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ArticleType } from "@/types/type";
-import BestArticle from "./BestArticle"; // BestArticle 컴포넌트의 경로에 따라 수정해야 할 수 있습니다.
+import BestArticle from "./BestArticle";
 import styles from "@/styles/BestArticle.module.css";
 
 interface BestArticleListProps {
