+++ pages/boards.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
 import Head from "next/head";
 import DropdownMenu from "@/components/DropdownMenu";
-import BestArticleBoards from "@/components/BestArticleBoards";
-import ArticleBoards from "@/components/ArticleBoards";
+import BestArticleBoards from "@/components/Boards/BestArticleBoards/BestArticleBoards";
+import ArticleBoards from "@/components/Boards/ArticleBoards/ArticleBoards";
 import styles from "@/styles/boards.module.css";
 
 //export async function getServerSideProps() {}
